replace spec test with unit test in tests/unit_node/http2_test.ts
