--- tests/specs/node/http2_allow_http1_fallback/__test__.jsonc
@@ -1,4 +1,0 @@
-{
-  "args": "run -A main.ts",
-  "output": "main.out"
-}
