--- tests/specs/node/http2_allow_http1_fallback/main.out
@@ -1,2 +1,0 @@
-request httpVersion: 1.1
-client response: 200 ok
